refactor: move transformation definition DSL into `dts-core`
