+++ Cargo.toml
@@ -21,7 +21,6 @@ clap_generate = "= 3.0.0-rc.7"
 dts-core = { path = "crates/dts-core" }
 dts-json = { path = "crates/dts-json" }
 grep-cli = "0.1.6"
-indexmap = "1.7.0"
 indoc = "1.0"
 rayon = "1.5.1"
 regex = "1.5.4"
