+++ crates/dts-core/Cargo.toml
@@ -8,6 +8,7 @@ csv = "1.1.6"
 glob = "0.3.0"
 hcl = { path = "../hcl" }
 dts-json = { path = "../dts-json" }
+indexmap = "1.7.0"
 json5 = "0.4.1"
 jsonpath-rust = "0.1.3"
 pathdiff = "0.2.1"
