+++ crates/dts-core/src/lib.rs
@@ -6,7 +6,6 @@ use std::path::{Path, PathBuf};
 
 pub use encoding::*;
 pub use error::*;
-pub use parsers::func_sig;
 pub use sink::Sink;
 pub use source::Source;
