+++ crates/dts-core/src/parsers/mod.rs
@@ -1,6 +1,5 @@
 mod error;
 pub mod flat_key;
-#[allow(missing_docs)]
 pub mod func_sig;
 pub mod gron;
