+++ crates/dts-core/src/transform/mod.rs
@@ -1,9 +1,11 @@
 //! Data transformation utilities.
 
+pub mod definition;
 mod error;
 pub(crate) mod key;
 pub(crate) mod sort;
 
+pub use definition::{Arg, Definition, DefinitionMatch, Definitions};
 pub use error::*;
 
 use crate::parsers::flat_key::{self, KeyPart, KeyParts};
